igmp: make an unset timer ID unambiguous, and free the whole source list
The al_timerid, al_query and al_versiontimer fields hold timer IDs, which
timer_set() hands out as a positive int, but they were declared uint32_t
and u_long.  A failed SetTimer() stored its -1 as 0xffffffff, which every
"if (g->al_timerid)" test reads as a running timer, so the group looks
armed while nothing will ever expire it.  Declare them int and let the
three Set*Timer() wrappers return 0, which timer_set() reserves and
timer_clear() ignores, so testing the field means what it looks like it
means.

DelVif() freed only the head of the group's source list.  Not reachable
today, because only SSM groups have sources and that path returns
earlier, but stop_vif() walks the list to free it and the two should not
disagree about who owns those entries.

+++ src/vif.h
@@ -285,10 +285,10 @@ struct listaddr {
     uint8_t	     al_mv;		/* router mrouted version	    */
     uint8_t	     al_old;            /* time since heard old report      */
     uint8_t	     al_index;		/* neighbor index		    */
-    uint32_t	     al_timerid;        /* timer for group membership	    */
-    uint32_t	     al_query;		/* timer for repeated leave query   */
+    int		     al_timerid;        /* timer for group membership	    */
+    int		     al_query;		/* timer for repeated leave query   */
     uint16_t	     al_flags;		/* flags related to this neighbor   */
-    u_long	     al_versiontimer;	/* timer for version switch         */
+    int		     al_versiontimer;	/* timer for version switch         */
 };
 #define	al_genid	al_alu.alu_genid
 #define	al_reporter	al_alu.alu_reporter
